fixed file structure and file path names in fish
