--- scripts/FishDataProvider.js
@@ -1,14 +1,0 @@
-const fishCollection = [
-    {
-        name: "Bart",
-        species: "fish",
-        location: "",
-        length: "",
-        diet: "",
-        food: "crustaceans",
-    }
-]
-
-export const useFish = () => {
-    return fishCollection.slice()
-}
