+++ scripts/main.js
@@ -1,5 +1,5 @@
-import { useFish } from './FishDataProvider.js'
-import { FishList } from './FishList.js'
+import { useFish } from './fish/FishDataProvider.js'
+import { FishList } from './fish/FishList.js'
 
 const allTheFish = useFish()
